requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
aiosqlite>=0.19.0
pyyaml>=6.0.1
aiofiles>=23.2.1
python-multipart>=0.0.6
reportlab>=4.0.0
httpx>=0.25.0
